⚗️ Add WIP on a software component plugin system
This system provides the means to aggregate commands by type as well as
the existing `manage` command group.

SoftwareComponents implement an abstract base class and register with a
SoftwareComponentManager, which is used by the CLI to list available
commands and components.

A dummy software component illustrates how the system is intended to be
used.

Integration with the concept of actions and the software component
scanning is outstanding.

* Depend on development version of rich-click

Relates to #1