🧱 Use loguru as logging framework
The logging module in the stdlib does not allow switching the log level
at runtime. It is also rather complicated to configure for what it is
supposed to do. Adding a new log level, for example, is non-trivial.

Loguru is a simple logging framework that already includes the TRACE log
level but also allows adding arbitrary ones. It plays nicely with rich
and is easy to configure.

Compared with competitors it has the most stars on GitHub, it is
actively maintained, and its documentation is okay. Sadly, there is no
doocumentation in VS code. But the online documentation is fine for
configuring it once. Afterwards, it should be simple to use so there is
no requirement for great documentation in the IDE.

* Add a basic theme to modify log colors

Relates to #15
Closes #15