🧱 Add buildah integration util
By using buildah it is possible to build container images without having
to create a Dockerfile. This is paving the way for being able to build
environment containers.

The implemented unit tests aren't really unit tests as they run real
commands. Either they have to be marked as integration tests or the
subprocess call has to be mocked.

Relates to #7