✨ Add option to deploy environment to container
This feature is currently limited to systems using the apt package
manager, i.e., ubuntu. The mechanism used for deploying to a container
is not yet final.

If it is kept then a large improvement would be to provide base images
for supported Linux derivates that already include a version of dagos
as installing it first is a large time sink. This is already painfully
noticable in the related BDD test.